feat(skill): render and page `skill show` in a terminal
`daft skill show` now renders the embedded SKILL.md with daft's markdown
skin and pages it (like `daft release-notes`) when stdout is a terminal,
so a human can read all ~860 lines comfortably. Piped or redirected it
still emits the raw source, byte-identical to what `daft skill install`
writes — the `> .../SKILL.md` manual-install path, `| head`, and agent
consumption are unchanged (TTY-gated: non-terminal always gets raw
bytes). `--no-pager` prints the rendered skill without a pager.

Extract the terminal markdown renderer (daft skin + OSC 8 link
rewriting) out of `commands::release_notes` into a shared
`output::markdown` module so both commands share one implementation.

Regenerate the git-daft-skill-show man page and CLI doc for the new
flag and about text; refresh the hand-authored daft-skill-show CLI page
and the test plan's Show section.

Part of #664.
