+++ src/commands/release_notes.rs
@@ -4,11 +4,8 @@
 /// using the system pager (like git does).
 use anyhow::Result;
 use clap::Parser;
-use regex::Regex;
 use serde::Serialize;
 use std::io::{self, IsTerminal};
-use termimad::MadSkin;
-use termimad::crossterm::style::Color;
 
 use crate::output::emit::{self, EmitArgs, EmitPayload, Format};
 
@@ -301,73 +298,12 @@ fn output_full(releases: &[Release], no_pager: bool) -> Result<()> {
 
     // Render markdown if outputting to a terminal
     let output = if io::stdout().is_terminal() {
-        render_markdown(&markdown)
+        crate::output::markdown::render(&markdown)
     } else {
         markdown
     };
 
     display_with_pager(&output, no_pager)
-}
-
-/// Render markdown to terminal-formatted text with colors
-fn render_markdown(markdown: &str) -> String {
-    let processed = markdown_links_to_osc8(markdown);
-    let skin = create_daft_skin();
-    skin.term_text(&processed).to_string()
-}
-
-/// Convert markdown links `[text](url)` to OSC 8 terminal hyperlinks.
-///
-/// OSC 8 is an escape sequence supported by modern terminals (iTerm2, Kitty,
-/// GNOME Terminal, Windows Terminal, WezTerm, etc.) that makes text clickable.
-/// Terminals that don't support it simply display the link text, which is a
-/// graceful degradation from the raw markdown syntax.
-fn markdown_links_to_osc8(markdown: &str) -> String {
-    let link_re = Regex::new(r"\[([^\]]+)\]\(([^)]+)\)").expect("valid regex");
-    link_re
-        .replace_all(markdown, "\x1b]8;;$2\x1b\\\x1b[34m$1\x1b[39m\x1b]8;;\x1b\\")
-        .into_owned()
-}
-
-/// Create a custom skin with daft's orange accent color
-fn create_daft_skin() -> MadSkin {
-    // Daft orange accent color
-    let orange = Color::Rgb {
-        r: 255,
-        g: 140,
-        b: 0,
-    };
-    let light_orange = Color::Rgb {
-        r: 255,
-        g: 180,
-        b: 100,
-    };
-
-    let mut skin = MadSkin::default();
-
-    // Headers in orange
-    skin.set_headers_fg(orange);
-
-    // Bold text in orange
-    skin.bold.set_fg(orange);
-
-    // Italic in a lighter orange
-    skin.italic.set_fg(light_orange);
-
-    // Inline code with subtle styling
-    skin.inline_code.set_fg(Color::Rgb {
-        r: 200,
-        g: 200,
-        b: 200,
-    });
-
-    // Bullet points in orange
-    skin.bullet.set_fg(orange);
-
-    // Horizontal rules in orange
-    skin.horizontal_rule.set_fg(orange);
-
-    skin
 }
 
 /// Display content using pager if appropriate.
@@ -450,49 +386,5 @@ mod tests {
         let releases = parse_changelog(content).unwrap();
         assert_eq!(releases.len(), 1);
         assert_eq!(releases[0].version, "1.0.0");
-    }
-
-    #[test]
-    fn test_render_markdown_produces_ansi() {
-        let md = "## Heading\n\n**bold** text";
-        let rendered = render_markdown(md);
-        // Rendered output should contain ANSI escape codes (start with \x1b[)
-        assert!(
-            rendered.contains("\x1b["),
-            "Rendered markdown should contain ANSI codes"
-        );
-    }
-
-    #[test]
-    fn test_markdown_links_to_osc8() {
-        let input = "see [#42](https://github.com/org/repo/pull/42) for details";
-        let result = markdown_links_to_osc8(input);
-        assert!(
-            result.contains("\x1b]8;;https://github.com/org/repo/pull/42\x1b\\"),
-            "should contain OSC 8 open sequence with URL"
-        );
-        assert!(
-            result.contains("\x1b[34m#42\x1b[39m"),
-            "should contain blue-colored link text"
-        );
-        assert!(
-            !result.contains("[#42]"),
-            "should not contain raw markdown link syntax"
-        );
-    }
-
-    #[test]
-    fn test_markdown_links_to_osc8_multiple() {
-        let input = "[a](https://a.com) and [b](https://b.com)";
-        let result = markdown_links_to_osc8(input);
-        assert!(result.contains("\x1b]8;;https://a.com\x1b\\\x1b[34ma\x1b[39m\x1b]8;;\x1b\\"));
-        assert!(result.contains("\x1b]8;;https://b.com\x1b\\\x1b[34mb\x1b[39m\x1b]8;;\x1b\\"));
-    }
-
-    #[test]
-    fn test_markdown_links_to_osc8_no_links() {
-        let input = "plain text with no links";
-        let result = markdown_links_to_osc8(input);
-        assert_eq!(result, input);
     }
 }
